Add Base class and split off HTTP into adapter
Our "base" file/entry was `HTTPClient` which wasn't great. This renames
that to a more generic `Base` class and adds the concept of Adapters.

An adapter is responsible for creating and sending requests, then
returning an instance of `Response`. This would let people use other
HTTP clients or even non HTTP transports.