Expose errors and extensions
GraphQL responses can have 3 possible root keys:

* `data`
* `errors`
* `extensions`

This exposes `errors` and `extensions` as arrays.

Previously we also raised an exception whenever `errors` was present,
but this wasn't a good default since execution errors can occur and data
can still be returned.

Now a `ResponseError` is only raised if `errors` exist and `data` does
not (or is null).