Add basic response object deserialization
Previously deserialization was tied to the `query` method and would only
worked alongside a Query object. This left `raw_query` without any nice
deserialization and method access of responses.

This adds a new `ResponseObject` which is similar to `GraphObject` but
skips anything to do with query objects. It simply provides nice method
access from response data.